Separate out StatusHeader from user.js.
