+++ public/record2/components/user.js
@@ -31,41 +31,6 @@ var tabs = [
     { path: 'comment',  name: 'コメント',     handler: CommentViewModule.commentView },
 ];
 
-var StatusHeader = React.createClass({
-    mixins: [Router.State],
-
-    getDefaultProps: function() {
-        return {
-            toolBar: function() {
-                return null;
-            }
-        };
-    },
-
-    render: function() {
-        var _tabs = tabs.map(function(tab) {
-            var className = 'status_tabbar_button';
-            if (tab.path === this.props.tabName) {
-                className += ' selected';
-            }
-            return (
-                    <li className={className} key={tab.path}>
-                    <Link to={tab.path} params={{
-                        token: this.getParams().token,
-                        report: this.getParams().report,
-                    }}>{tab.name}</Link>
-                    </li>
-            );
-        }.bind(this));
-        return (
-                <div className="status_header">
-                {this.props.toolBar()}
-                <ul className="status_tabbar">{_tabs}</ul>
-                </div>
-        );
-    }
-});
-
 var routes = tabs.map(function(tab) {
     return (
             <Route name={tab.path} path={tab.path} handler={tab.handler} key={tab.path}/>
@@ -81,5 +46,4 @@ var UserRoute = (
 
 module.exports = {
     userRoute: UserRoute,
-    statusHeader: StatusHeader
 };
